refactor: place lock in RustyArchivalMutatorSet
This is a weak or partial encapsulation of AtomicRw lock inside
RustyArchivalMutatorSet.  The lock type is changed from a std
RwLock to a tokio RwLock.

We expose the internal lock publicly and In fact most locking is still
performed by third parties.

So this change is not much of a win by itself, but it is more
consistent with the other locking encapsulation refactors and should
be helpful if/when we use async database in twenty_first::storage.