fix: send syncing start/stop messages to miner
The mining loop is now occurring in a rayon parallel loop
meaning it runs on multiple threads.

Thus it can no longer await async futures.  So we cannot check
the GlobalState syncing flag, which requires an async lock
acquisition.

Instead, the main task sends StartSyncing and StopSyncing
messages to the mining task which cancels the mining thread
when it receives StartSyncing.