feat: generate proofs out-of-process
invokes external process `triton-vm-prover` to
generate each proof.

this is an implementeation detail of ConsensusProgramProverJob.

ConsensusProgramProverJob is now an async job, and when it is
aborted the external process is also killed.

presently triton-vm-prover must exist in $PATH.

A future commit will do something better in terms of executing the
correct executable to match neptune-core (debug, release).

Also the external process is not presently invoked when running
unit tests.