feat: job cancellation of JobQueue jobs
This is is a significant upgrade to the JobQueue.

JobQueue::add_job() now returns a JobHandle which can be
used to cancel() a job or complete() a job, or wait for a
job result().

Job cancellation works for both async and sync jobs. It is automatic
for async jobs, but requires some cooperation from sync jobs.  Sync
jobs should check a job-cancellation channel regularly and abort if
necessary.

changes:
 * defines "real" thiserror errors, instead of anyhow errors.
 * JobHandle::result() replaces JobQueue::add_and_await_result().
 * adds cancel_job tests for sync and async jobs.
 * improved doc-comments
 * test/example job now demonstrates cancellation for sync jobs.