refactor(`PrimitiveWitness`): Time-lock half of coinbase
In the main code path for arbitrary instances of `PrimitiveWitness`,
the `num_inputs` now has type `Option<usize>` so that it either counts
the number of input UTXOs in the transaction, or else indicates that
the generated `PrimitiveWitness` is one for a coinbase transaction
(*i.e.*, one with a set coinbase amount). When a `PrimitiveWitness`
for a coinbase transaction is generated then there are 2x as many
output UTXOs and half of them are time-locked for 3 years.