fix: store known keys in HashSet to ensure unique
Implements Eq and std::hash::Hash for SpendingKey so it can be stored
in a HashMap, HashSet.

Fixes a problem where the first generation address (index 0) can be
stored twice.

The root of this problem is actually a workaround for tests that do
not correctly call next_unused_spending_key(). A future commit will
fix the tests and remove the workaround, which should remove the dup.

It will still be useful to be able to store keys in a hash based
collection, for 3rd parties also. Using a HashSet in the wallet
provides a nice guarante.

Arguably once a test is in place that no dups occur, we could go
back to using a Vec which is a bit faster.