fix: Test `upgrade_proof_collection_to_single_proof_foreign_tx`
The test in question made an assumption that is no longer valid,
namely that the TXID of a transaction does not change if its proof
is upgraded. The reason why this is no longer valid is because the
proof upgrader claims a fee for his work; and exacts that fee by
merging in another transaction that has a negative fee.