refactor: remove data_dir from GlobalStateLock
addresses some review comments in #334

moves GlobalStateLock::data_dir to NeptuneRPCServer::data_directory

This is done because there isn't any clear need for this field in
GlobalStateLock right now, and to avoid cluttering up that struct
unnecessarily.

field is renamed to data_directory to distinguish from
CliArg::data_dir which is a different data type and path.