fix: report guessing time correctly
fixes #405

Refactors mining loop a bit such that:
1. state transition from not-guessing to guessing is identifiable.
2. state transition from guessing to inactive is identifiable.
3. globalState mining_status is set to guessing when guessing starts
4. ming_status is set to inactive when guessing starts.

Previously mining_status was being set to guessing when guessing first
starts, then inactive after 20 seconds, then guessing again, then
inactive after 20 seconds, and so on until a block is found or guessing
is cancelled.  Thus the max reported duration of guessing was 20
seconds.

With this change, the reported duration accurately reflects the actual
amount of time spent in the guessing state, regardless of the restarts
every 20 seconds.

There is also a minor cleanup that aims to preserve existing logic but
move duplicated code into a common place, which makes this complicated
fn a bit more readable.