feat: handle panics in job-queue jobs
closes #524

The job-queue now detects panics returned as errors from the tokio JoinHandle
representing the job tasks.

After a job panics, the job-queue continues processing jobs as normal. However
the calling code may detect the error and decide to stop the queue  or even
shutdown the process.

JobCompletion has a new variant Panicked that holds panic data.
JobHandleError has a new variant JobPanicked that holds panic data.

Unfortunately panic data is not `Sync` and cannot be sent between threads.
To work-around this, JobHandleErrorSync has been added so that callers can
easily convert JobHandleError into a type that is Sync because the panic data
is transformed into a String.  This provides convenience if desired while
preserving the original panic data in return types.

new tests:

* panic_in_async_job_ends_job_cleanly
* panic_in_blocking_job_ends_job_cleanly